add send_letters to add competition
