+++ adaptive_hockey_federation/competitions/views.py
@@ -27,6 +27,7 @@ from django.views.generic.edit import (
 from django.views.generic.list import ListView
 from main.controllers.team_views import CityListMixin
 from main.controllers.utils import get_team_href
+from users.utilits.send_mails import send_email
 
 
 class CompetitionListView(
@@ -152,6 +153,11 @@ class AddTeamToCompetition(
             )
             team = get_object_or_404(Team, id=kwargs["pk"])
             competition.teams.add(team)
+            send_email(
+                instance=team,
+                competition=competition,
+                type=True,
+            )
         return super(AddTeamToCompetition, self).dispatch(
             request, kwargs["competition_id"]
         )
