+++ adaptive_hockey_federation/templates/emailing/email.html
@@ -2,6 +2,7 @@
 <html xmlns="http://www.w3.org/1999/xhtml" xmlns:v="urn:schemas-microsoft-com:vml"
     xmlns:o="urn:schemas-microsoft-com:office:office">
 <head>
+    {% load static %}
     <title>
     </title>
     <!--[if !mso]><!-- -->
@@ -9,6 +10,7 @@
     <!--<![endif]-->
     <meta http-equiv="Content-Type" content="text/html; charset=UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1">
+    <link rel="stylesheet" href="{% static 'css/base_style.css' %}">
     <style type="text/css">
         #outlook a {
             padding: 0;
@@ -67,6 +69,223 @@
     </style>
     <style type="text/css">
     </style>
+
+    <style type="text/css">
+        .welcome_container {
+          text-align: center;
+          width: 100%; 
+          padding: 0px;
+          border: 1px solid #340061; 
+          border-radius: 10px;
+          background: url('https://i.postimg.cc/KjFtxKnG/fone.png');
+        }
+      
+        .welcome_header {
+          width: 100%;
+          height: 30%;
+          display: grid;
+          grid-template-columns: repeat(12, 1fr);
+          justify-content: space-between;
+          align-items: center;
+          background-color: white;
+          border-radius: 10px;
+        }
+      
+        .welcome-logo {
+          margin: 0 5px 0 15px;
+          position: left;
+          max-width: 100px;
+          max-height: 100px;
+          grid-column: 1 / 2;
+          z-index: 3;
+        }
+      
+        @media screen and (min-width: 768px) and (max-width: 1280px) {
+          .welcome-logo {
+            margin: 0 4px 0 10px;
+            max-width: 80px;
+            max-height: 80px;
+          }
+        }
+      
+        @media screen and (min-width: 320px) and (max-width: 767px) {
+          .welcome-logo {
+            margin: 0 3px 0 6px;
+            max-width: 40px;
+            max-height: 40px;
+          }
+        }
+      
+        .welcome-header-text {
+          position: relative;
+          margin: 0;
+          padding: 0;
+          color: #340061;
+          grid-column: 2 / span 4;
+          grid-row: 1;
+          text-decoration: none;
+          font-family: sans-serif;
+          font-size: 18px;
+          font-weight: 700;
+          font-weight: bold;
+          z-index: 5;
+        }
+      
+        @media screen and (min-width: 1441px) and (max-width: 1920px) {
+          .welcome-header-text {
+            font-size: 16px;
+          }
+        }
+      
+        @media screen and (min-width: 1281px) and (max-width: 1440px) {
+          .welcome-header-text {
+            grid-column: 2 / span 4;
+            display: flex;
+            justify-content: center;
+          }
+        }
+      
+        @media screen and (min-width: 768px) and (max-width: 1280px) {
+          .welcome-header-text {
+            font-size: 8px;
+          }
+        }
+      
+        @media screen and (min-width: 400px) and (max-width: 767px) {
+          .welcome-header-text {
+            font-size: 6px;
+          }
+        }
+      
+        @media screen and (min-width: 320px) and (max-width: 399px) {
+          .welcome-header-text {
+            font-size: 4px;
+          }
+        }
+      
+      
+        .welcome-rect-container {
+          position: relative;
+          display: block;
+          min-height: 100%;
+          grid-column: 4 / -1;
+          grid-row: 1;
+          border-radius: 10px;
+        }
+      
+        @media screen and (min-width: 1441px) and (max-width: 1920px) {
+          .welcome-rect-container {
+            grid-column: 6 / -1;
+          }
+        }
+      
+        .welcome-ellipse {
+          position: absolute;
+          top : 0;
+          left: -45%;
+          min-width: 70%;
+          height: 100%;
+          background-color: white;
+          border-radius: 9999px;
+          z-index: 2;
+        }
+      
+        @media screen and (min-width: 768px) and (max-width: 1280px) {
+          .welcome-ellipse {
+            left: -45%;
+          }
+        }
+      
+        @media screen and (min-width: 320px) and (max-width: 767px) {
+          .welcome-ellipse {
+            left: -55%;
+          }
+        }
+      
+        .welcome-rect-container {
+          position: relative;
+          display: block;
+          min-height: 100%;
+          grid-column: 4 / -1;
+          grid-row: 1;
+        }
+      
+        @media screen and (min-width: 1441px) and (max-width: 1920px) {
+          .welcome-rect-container {
+            grid-column: 4 / -1;
+          }
+        }
+      
+        .welcome-rect-container {
+          position: relative;
+          max-width: 100%;
+          min-height: 150px;
+          z-index: 1;
+        }
+      
+        @media screen and (min-width: 768px) and (max-width: 1280px) {
+          .welcome-rect-container {
+            min-height: 90px;
+          }
+        }
+      
+        @media screen and (min-width: 320px) and (max-width: 767px) {
+          .welcome-rect-container {
+            min-height: 6%;
+            max-height: 60px;
+          }
+        }
+        .welcome_footer a {
+          position: relative;
+          color: white;
+          text-decoration: none;
+          height: 100px;
+          font-weight: 400;
+          padding: 3px;
+          margin-top: 2px;
+          background-color: #2642bd;
+          border-radius: 8px;
+      
+        }
+      
+        .welcome_footer a:hover {
+          color: white;
+          background-color: #340061;
+        }
+      
+        .welcome_footer {
+          color: black;
+          text-decoration: none;
+          width: 100%;
+          height: 15%;
+          background: url('https://i.postimg.cc/x1qMfVbz/footer.png');
+          max-width: 100vw;
+          border-radius: 10px;
+        }
+      
+        .welcome_footer p {
+          font-family: sans-serif;
+          color: #340061;
+          font-weight: 700;
+          padding: 10px;
+          margin: 0;
+          border-radius: 10px;
+        }
+      
+        .welcome_body_list {
+          list-style-type: none; 
+          text-align: left;
+          font-family: sans-serif;
+          color: #340061;
+          font-weight: 700;
+          font-weight: bold;
+        }
+      
+        .card-img-top {
+          border-radius: 50%;
+          margin-left: -250px;
+        }
+      </style>
 </head>
 <body style="background-color:#f9f9f9;">
